fix : get update content
