+++ src/hooks/index.ts
@@ -3,7 +3,5 @@ export * from './useDetectResolution';
 export * from './useHandleOpenModal';
 export * from './useViewportWidth';
 export * from './useArticleElement';
-export * from './article/useResolveSaveArticleHandler';
-export * from './article/useHandleDelete';
 export * from './detail-page/useSettingDetailPage';
-export * from './article/useSettingListPage';
+export * from './article';
